BREAKING!: remove private: true from packages - re-export BoardSetupFunction from react-board
