+++ packages/board-plugins/package.json
@@ -15,7 +15,6 @@
   },
   "repository": "https://github.com/wixplosives/wcs-core/tree/master/packages/board-plugins",
   "homepage": "https://github.com/wixplosives/wcs-core",
-  "private": true,
   "publishConfig": {
     "access": "public"
   }
