+++ packages/react-board/src/index.ts
@@ -1,1 +1,2 @@
 export * from './create-board';
+export * from './types';
